\name{RM2006-package}
\alias{RM2006-package}
\docType{package}
\title{
\packageTitle{RM2006}
}
\description{
\packageDescription{RM2006}
}
\author{
\packageAuthor{RM2006}

Maintainer: \packageMaintainer{RM2006}
}
\references{
Zumbach, G. (2007) The Riskmetrics 2006 methodology. Available at SSRN: https://ssrn.com/abstract=1420185 or http://dx.doi.org/10.2139/ssrn.1420185
}

\keyword{Risk Metrics}
\keyword{Conditional Covariance Matrix}


